financebro-backend/requirements.txt
brotli
django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-ninja==0.21.0
django-solo==2.0.0
dj-rest-auth==3.0.0
lxml==4.9.2
orjson==3.8.6
python-dateutil==2.8.2
requests==2.28.2
sentry-sdk==1.20.0
stripe
whitenoise==6.4.0
psycopg2==2.9.6
git+https://github.com/sendinblue/APIv3-python-library.git
